+++ onsiteprint-plugin/assets/js/inc/template/get-template.js
@@ -1,8 +1,8 @@
 /* ------------------------------------------------------------------------
  #  JS Part Name: Get Template
  *  Getting the Template from the Local Storage.
- ?  Updated: 2025-06-01 - 18:31 (Y:m:d - H:i)
- ?  Info: Added new Get Template Script with new Function, opGetTemplate().
+ ?  Updated: 2025-06-01 - 18:35 (Y:m:d - H:i)
+ ?  Info: Added the Local Storage Data and the Filter of the Template List.
 ---------------------------------------------------------------------------
  #  TABLE OF CONTENTS:
 ---------------------------------------------------------------------------
@@ -15,6 +15,7 @@
  #  1. Import Functions from Scripts
 --------------------------------------------------------------------------- */
 import * as opModuleBasic from '../basic.js'
+import { opGetLocalStorageData } from '../api/get-local-storage-data.js'
 
 /* ------------------------------------------------------------------------
  #  2. Function: Get Template from the Local Storage.
@@ -30,6 +31,44 @@ export async function opGetTemplate( debug, templateId ) {
         ////* Set the Parameter If is not defined (true or false).
         if ( debug !== true ) debug = false
         if ( debug ) console.group( `${ functionName }()` )
+
+        ///// Get Local Storage Data.
+        const localStorageData = await opGetLocalStorageData( debug, 'TEMPLATES' )
+        
+        ///// Validate the Response from the Local Storage Data.
+        if ( localStorageData.error !== false ) throw opModuleBasic.opReturnResponse( true, 400, { 
+            message: `Something went wrong getting the Local Storage Data!`,
+            line: opModuleBasic.errorLine(),
+            function: functionName
+        } )
+
+        ///// Get Template List.
+        const templateList = localStorageData.response.details.templateList
+
+        ///// Validate Template List.
+        if ( ! templateList || ! templateList[0] ) throw opModuleBasic.opReturnResponse( true, 404, { 
+            message: `No Templates have been created yet!`,
+            line: opModuleBasic.errorLine(),
+            function: functionName
+        } )
+
+        ///// Find the Template.
+        let templateItem = templateList.filter( templateItem => templateItem.templateCreationDate === Number( templateId ) )
+
+        ///// Validate Template Item.
+        if ( ! templateItem[0] ) throw opModuleBasic.opReturnResponse( true, 404, { 
+            message: `No Template was found!`,
+            line: opModuleBasic.errorLine(),
+            function: functionName
+        } )
+
+        ///// Return the Response.
+        return opModuleBasic.opReturnResponse( false, 200, { 
+            message: `The Template was Found!`, 
+            line: opModuleBasic.errorLine(),
+            function: functionName,
+            details: templateItem[0]
+        }, debug )
 
     } catch( errorResponse ) {